using np.where to find the place of the `linebreak` and np.split to cut at these positions is a little bit faster and needs less lines
In principle the result of the np.split call is already the `time_lines`,
but each time_line still contains the `linebreak` in the end.
so when the time_line is [35, 105] the `part` is [35, 105, 255].

Luckily I still need to convert from np.array([35, 105], dtype=np.uint8)
to an `array('B')`, within this convertsion I also split off the last element.